Move definition of EventTableSubset into separate file
